fix business rules to int value

+++ src/main/java/io/github/emanuelcerqueira/ifoodbackendchallenge/core/services/ClassicalGenreSuggestionRule.java
@@ -13,6 +13,6 @@ public class ClassicalGenreSuggestionRule implements GenreSuggestionRule {
 
     @Override
     public boolean isTemperatureSuitableToMusicGenre(Double temperature) {
-        return temperature < 10;
+        return temperature.intValue() < 10;
     }
 }
